+++ controlplane/dashboard/package.json
@@ -26,7 +26,7 @@
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^7.0.0",
-    "naive-ui": "^2.23.1",
-    "vfonts": "^0.1.0"
+    "naive-ui": "2.28.2",
+    "vfonts": "^0.0.3"
   }
 }
