+++ controlplane/dashboard/src/main.js
@@ -3,7 +3,7 @@ import App from './App.vue'
 import router from './router'
 import naive from 'naive-ui'
 import store from './store/store'
-
+ 
 let app = createApp(App)
     app.use(router).
     use(store).
